+++ packages/dsl-java/src/main/java/io/contract_testing/contractcase/configuration/InteractionSetup.java
@@ -1,5 +1,6 @@
 package io.contract_testing.contractcase.configuration;
 
+import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.contract_testing.contractcase.exceptions.ContractCaseConfigurationError;
@@ -20,42 +21,105 @@ import org.jetbrains.annotations.NotNull;
  */
 public class InteractionSetup {
 
+  private static final ObjectMapper mapper = new ObjectMapper()
+          .configure(JsonParser.Feature.INCLUDE_SOURCE_IN_LOCATION, true);
+
   private final Map<String, String> stateVariables;
   private final Map<String, String> mockSetup;
   private final Map<String, InvokeCoreFunction> functions;
-  private final ObjectMapper mapper;
 
 
   private InteractionSetup(ConnectorSetupInfo connectorSetupInfo) {
     this.stateVariables = connectorSetupInfo.stateVariables();
     this.mockSetup = connectorSetupInfo.mockSetup();
     this.functions = connectorSetupInfo.functions();
-    this.mapper = new ObjectMapper();
   }
 
   /**
-   * Get a state variable set by a state definition or state handler
+   * Get a state variable set by a state definition or state handler. This method assumes that the variable is a string,
+   * which is not always true. For this reason, this method is deprecated, and will be private in a future release.
+   * Prefer using {@link #getStateVariable(String name, Class)}, {@link #getStringStateVariable(String name)}
+   * or {@link #getIntegerStateVariable(String name)} instead
+   *
+   * @param key the state variable key that was defined in the state handler
+   * @return The value of this state variable
+   * @throws ContractCaseConfigurationError if there is no value for this key
+   * @deprecated Prefer using
+   */
+  @Deprecated
+  public String getStateVariable(String key) {
+    return this.getStringStateVariable(key);
+  }
+
+  /**
+   * Get a state variable set by a state definition or state handler. This returns the raw string, which is json.
    *
    * @param key the state variable key that was defined in the state handler
    * @return The value of this state variable
    * @throws ContractCaseConfigurationError if there is no value for this key
    */
-  public String getStateVariable(String key) {
+  private String internalGetStateVariable(String key) {
     if (this.stateVariables.get(key) == null) {
       final var keys = new ArrayList<>(this.stateVariables.keySet());
       throw new ContractCaseConfigurationError(
-          "Can't get state variable '" + key + "', as it's not present in the SetupInfo"
-              + "'. Check the variable is defined in the contract. "
-              + (keys.size() == 0
-              ? "There are no currently defined state variables"
-              : "Currently defined variables are: \n"
-                  + keys
-                  .stream().map(s -> "    " + s)
-                  .collect(Collectors.joining("\n"))),
-          "BAD_INTERACTION_DEFINITION"
+              "Can't get state variable '" + key + "', as it's not present in the SetupInfo"
+                      + "'. Check the variable is defined in the contract. "
+                      + (keys.isEmpty()
+                      ? "There are no currently defined state variables"
+                      : "Currently defined variables are: \n"
+                        + keys
+                      .stream().map(s -> "    " + s)
+                      .collect(Collectors.joining("\n"))),
+              "BAD_INTERACTION_DEFINITION"
       );
     }
     return this.stateVariables.get(key);
+  }
+
+  /**
+   * Gets a state variable sets by a state definition or state handler. 
+   *
+   * @param name the name of the state variable
+   * @param type The type of the value
+   * @return the deserialised value
+   * @param <T> The type of the deserialised object.
+   * @throws ContractCaseConfigurationError if the value doesn't deserialise, or isn't found in the state variables.
+   */
+  public <T> T getStateVariable(String name, Class<T> type) {
+    var json = this.internalGetStateVariable(name);
+    try {
+      return mapper.readValue(json, type);
+    } catch(JsonProcessingException e ) {
+       throw new ContractCaseConfigurationError(
+          "Unable to deserialise Json parameter for '" + name + "'. Json was: " + json,
+           e
+       );
+    }
+  }
+
+  /**
+   * Convenience method to get an integer state variable set by a state definition or state handler
+   * Equivalent to calling {@code getStateVariable(key, Integer.class)}
+   *
+   * @param key the state variable key that was defined in the state handler
+   * @return The value of this state variable
+   * @throws ContractCaseConfigurationError if there is no value for this key
+   */
+  public Integer getIntegerStateVariable(String key) {
+    return this.getStateVariable(key, Integer.class);
+  }
+
+
+  /**
+   * Convenience method to get a string state variable set by a state definition or state handler.
+   * Equivalent to calling {@code getStateVariable(key, String.class)}
+   *
+   * @param key the state variable key that was defined in the state handler
+   * @return The value of this state variable
+   * @throws ContractCaseConfigurationError if there is no value for this key
+   */
+  public String getStringStateVariable(String key) {
+      return this.getStateVariable(key, String.class);
   }
 
   /**
@@ -132,5 +196,6 @@ public class InteractionSetup {
   public static InteractionSetup from(@NotNull ConnectorSetupInfo config) {
     return new InteractionSetup(config);
   }
+
 
 }