+++ packages/dsl-java/src/test/java/io/contract_testing/contractcase/test/function/verification/BrokenFunctionCallerVerificationTest.java
@@ -1,10 +1,12 @@
 package io.contract_testing.contractcase.test.function.verification;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.contract_testing.contractcase.ContractVerifier;
+import io.contract_testing.contractcase.VerificationResult;
 import io.contract_testing.contractcase.configuration.ContractCaseConfig.ContractCaseConfigBuilder;
 import io.contract_testing.contractcase.configuration.InvokableFunctions.InvokableFunction1;
 import io.contract_testing.contractcase.configuration.LogLevel;
@@ -14,7 +16,6 @@ import io.contract_testing.contractcase.exceptions.ContractCaseConfigurationErro
 import java.util.HashMap;
 import java.util.function.Function;
 import org.jetbrains.annotations.NotNull;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -22,12 +23,12 @@ public class BrokenFunctionCallerVerificationTest {
 
   static final ObjectMapper mapper = new ObjectMapper();
 
-  private ContractVerifier contract;
+  private ContractVerifier verifier;
 
 
   @BeforeEach
   void setup() {
-    contract = new ContractVerifier(ContractCaseConfigBuilder.aContractCaseConfig()
+    verifier = new ContractVerifier(ContractCaseConfigBuilder.aContractCaseConfig()
         .consumerName("Java Function Caller Example")
         .providerName("Java Function Implementer Example")
         .publish(PublishType.NEVER)
@@ -35,18 +36,18 @@ public class BrokenFunctionCallerVerificationTest {
         //      .logLevel(LogLevel.DEEP_MAINTAINER_DEBUG)
         .contractDir("verifiable-contracts")
         .build());
-    contract.registerFunction("NoArgFunction", () -> {
+    verifier.registerFunction("NoArgFunction", () -> {
       return null;
     });
 
-    contract.registerFunction(
+    verifier.registerFunction(
         "PageNumbers",
         convertJsonIntegerArg((Integer num) -> num + " pages")
     );
 
     var mockedStore = new HashMap<String, String>();
 
-    contract.registerFunction(
+    verifier.registerFunction(
         "keyValueStore",
         convertJsonStringArgs((String key) -> mockedStore.get(key))
     );
@@ -58,14 +59,14 @@ public class BrokenFunctionCallerVerificationTest {
 
     try {
       assertThrows(ContractCaseConfigurationError.class, () -> {
-        contract.runVerification(ContractCaseConfigBuilder.aContractCaseConfig()
+        verifier.runVerification(ContractCaseConfigBuilder.aContractCaseConfig()
             // Don't print any logs, otherwise CI logs get polluted with expected failures
             .logLevel(LogLevel.NONE)
             // Don't print results, otherwise CI logs get polluted with expected failures
             .printResults(false).throwOnFail(true).build());
       });
     } finally {
-      contract.close();
+      verifier.close();
     }
   }
 
@@ -76,7 +77,7 @@ public class BrokenFunctionCallerVerificationTest {
 
       // This contract verification should fail, but we'll ignore it, because
       // throwOnFail is set to false
-      contract.runVerification(ContractCaseConfigBuilder.aContractCaseConfig()
+      var contractHandles = verifier.prepareVerifications(ContractCaseConfigBuilder.aContractCaseConfig()
           // Don't print any logs, otherwise CI logs get polluted with expected failures
           .logLevel(LogLevel.NONE)
           // Don't print results, otherwise CI logs get polluted with expected failures
@@ -87,8 +88,22 @@ public class BrokenFunctionCallerVerificationTest {
           ).stateHandler("The map is not null", StateHandler.setupFunction(() -> {}))
           .stateHandler("The key 'foo' is set to 'bar'", StateHandler.setupFunction(() -> {}))
           .throwOnFail(false).build());
+
+      assertThat(contractHandles).isNotEmpty();
+      for (var contract : contractHandles) {
+        contract.testHandles().forEach(verifier::runPreparedTest);
+        var result = verifier.closePreparedVerification(contract);
+        assertThat(result.consumerSlug()).isEqualTo("java-function-caller-example");
+        assertThat(result.providerSlug()).isEqualTo("java-function-implementer-example");
+        assertThat(result.description().consumerName()).isEqualTo("Java Function Caller Example");
+        assertThat(result.description().providerName()).isEqualTo("Java Function Implementer Example");
+
+        assertThat(result.contractPath()).isNotNull();
+        assertThat(result.metadata().get("_case.version")).isNotNull();
+        assertThat(result.compatibility()).isEqualTo(VerificationResult.Compatibility.INCOMPATIBLE);
+      }
     } finally {
-      contract.close();
+      verifier.close();
     }
   }