+++ packages/case-definition-generator/CHANGELOG.md
@@ -1,5 +1,45 @@
 # Changelog
 
+## [0.29.1](https://github.com/case-contract-testing/contract-case/compare/@contract-case/case-definition-generator-v0.29.0...@contract-case/case-definition-generator-v0.29.1) (2026-06-04)
+
+
+### Miscellaneous Chores
+
+* **@contract-case/case-definition-generator:** Synchronize ContractCase versions
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @contract-case/case-plugin-base bumped from 0.29.0 to 0.29.1
+    * @contract-case/case-core-plugin-http bumped from 0.29.0 to 0.29.1
+    * @contract-case/case-core-plugin-function bumped from 0.29.0 to 0.29.1
+    * @contract-case/case-plugin-dsl-types bumped from 0.29.0 to 0.29.1
+  * devDependencies
+    * @contract-case/case-maintainer-config bumped from 0.29.0 to 0.29.1
+    * @contract-case/eslint-config-case-maintainer bumped from 0.29.0 to 0.29.1
+
+## [0.29.0](https://github.com/case-contract-testing/contract-case/compare/@contract-case/case-definition-generator-v0.28.2...@contract-case/case-definition-generator-v0.29.0) (2026-05-31)
+
+
+### Miscellaneous Chores
+
+* **@contract-case/case-definition-generator:** Synchronize ContractCase versions
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @contract-case/case-plugin-base bumped from 0.28.2 to 0.29.0
+    * @contract-case/case-core-plugin-http bumped from 0.28.2 to 0.29.0
+    * @contract-case/case-core-plugin-function bumped from 0.28.2 to 0.29.0
+    * @contract-case/case-plugin-dsl-types bumped from 0.28.2 to 0.29.0
+  * devDependencies
+    * @contract-case/case-maintainer-config bumped from 0.28.2 to 0.29.0
+    * @contract-case/eslint-config-case-maintainer bumped from 0.28.2 to 0.29.0
+
 ## [0.28.2](https://github.com/case-contract-testing/contract-case/compare/@contract-case/case-definition-generator-v0.28.1...@contract-case/case-definition-generator-v0.28.2) (2026-02-12)