+++ packages/case-plugin-base/CHANGELOG.md
@@ -1,5 +1,48 @@
 # Changelog
 
+## [0.29.1](https://github.com/case-contract-testing/contract-case/compare/@contract-case/case-plugin-base-v0.29.0...@contract-case/case-plugin-base-v0.29.1) (2026-06-04)
+
+
+### Miscellaneous Chores
+
+* **@contract-case/case-plugin-base:** Synchronize ContractCase versions
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @contract-case/case-plugin-dsl-types bumped from 0.29.0 to 0.29.1
+  * devDependencies
+    * @contract-case/case-maintainer-config bumped from 0.29.0 to 0.29.1
+    * @contract-case/eslint-config-case-maintainer bumped from 0.29.0 to 0.29.1
+
+## [0.29.0](https://github.com/case-contract-testing/contract-case/compare/@contract-case/case-plugin-base-v0.28.2...@contract-case/case-plugin-base-v0.29.0) (2026-05-31)
+
+
+### ⚠ BREAKING CHANGES
+
+* **plugin-dsl:** Change `describe` type to return structured `DescribeSegment` instead of string. Plugin authors will need to use the new factory functions to implement their describe methods instead of returning a string. This allows structured rendering of the describe result ([#1246](https://github.com/case-contract-testing/contract-case/issues/1246))
+
+### Features
+
+* Add error code NO_INTERACTIONS_DEFINED to help distinguish when there are no interactions in a contract definition ([6302c8e](https://github.com/case-contract-testing/contract-case/commit/6302c8e17f981b83413116e6ede279461625f161))
+
+
+### Code Refactoring
+
+* **plugin-dsl:** Change `describe` type to return structured `DescribeSegment` instead of string. Plugin authors will need to use the new factory functions to implement their describe methods instead of returning a string. This allows structured rendering of the describe result ([#1246](https://github.com/case-contract-testing/contract-case/issues/1246)) ([de9546b](https://github.com/case-contract-testing/contract-case/commit/de9546b468bcddbb371d02be3d2e8cc3712eeb42))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @contract-case/case-plugin-dsl-types bumped from 0.28.2 to 0.29.0
+  * devDependencies
+    * @contract-case/case-maintainer-config bumped from 0.28.2 to 0.29.0
+    * @contract-case/eslint-config-case-maintainer bumped from 0.28.2 to 0.29.0
+
 ## [0.28.2](https://github.com/case-contract-testing/contract-case/compare/@contract-case/case-plugin-base-v0.28.1...@contract-case/case-plugin-base-v0.28.2) (2026-02-12)