+++ packages/contract-case-dsl-js-jest/CHANGELOG.md
@@ -29,6 +29,52 @@
     * @contract-case/case-example-mock-types bumped from 0.14.0 to 0.15.1
     * @contract-case/cli bumped from 0.1.6 to 0.1.7
 
+## [0.29.1](https://github.com/case-contract-testing/contract-case/compare/@contract-case/contract-case-jest-v0.29.0...@contract-case/contract-case-jest-v0.29.1) (2026-06-04)
+
+
+### Bug Fixes
+
+* **java-dsl:** Add missing deserialisation functions for non-string state variables to the Java DSL. Previously, these didn't work at all. `getStateVariable(String)` is now deprecated, please use `getStateVariable(String, Class)` or the convenience functions getIntegerStateVariable and getStringStateVariable ([ffb18f8](https://github.com/case-contract-testing/contract-case/commit/ffb18f8294a68474b3e95fdf80d23b0418386a8d))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @contract-case/case-connector bumped from 0.29.0 to 0.29.1
+    * @contract-case/case-core-plugin-http-dsl bumped from 0.29.0 to 0.29.1
+    * @contract-case/case-definition-dsl bumped from 0.29.0 to 0.29.1
+    * @contract-case/case-entities-internal bumped from 0.29.0 to 0.29.1
+    * @contract-case/case-plugin-base bumped from 0.29.0 to 0.29.1
+    * @contract-case/case-plugin-dsl-types bumped from 0.29.0 to 0.29.1
+    * @contract-case/cli bumped from 0.29.0 to 0.29.1
+  * devDependencies
+    * @contract-case/cli bumped from 0.29.0 to 0.29.1
+    * @contract-case/eslint-config-case-maintainer bumped from 0.29.0 to 0.29.1
+
+## [0.29.0](https://github.com/case-contract-testing/contract-case/compare/@contract-case/contract-case-jest-v0.28.2...@contract-case/contract-case-jest-v0.29.0) (2026-05-31)
+
+
+### Miscellaneous Chores
+
+* **@contract-case/contract-case-jest:** Synchronize ContractCase versions
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @contract-case/case-connector bumped from 0.28.2 to 0.29.0
+    * @contract-case/case-core-plugin-http-dsl bumped from 0.28.2 to 0.29.0
+    * @contract-case/case-definition-dsl bumped from 0.28.2 to 0.29.0
+    * @contract-case/case-entities-internal bumped from 0.28.2 to 0.29.0
+    * @contract-case/case-plugin-base bumped from 0.28.2 to 0.29.0
+    * @contract-case/case-plugin-dsl-types bumped from 0.28.2 to 0.29.0
+    * @contract-case/cli bumped from 0.28.2 to 0.29.0
+  * devDependencies
+    * @contract-case/cli bumped from 0.28.2 to 0.29.0
+    * @contract-case/eslint-config-case-maintainer bumped from 0.28.2 to 0.29.0
+
 ## [0.28.2](https://github.com/case-contract-testing/contract-case/compare/@contract-case/contract-case-jest-v0.28.1...@contract-case/contract-case-jest-v0.28.2) (2026-02-12)